Enhance post detail page with authentication checks for liking and commenting features
